agent/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.1.8
langsmith>=0.1.78
llama-cpp-python>=0.2.90
huggingface-hub>=0.24.0
hf-transfer>=0.1.9
sentence-transformers>=3.0.1
faiss-cpu==1.11.0
pandas>=2.2.0
openpyxl>=3.1.5
pypdf>=4.2.0
python-docx>=1.1.2
numpy>=1.26.4
typer>=0.12.5
rich>=13.7.1
tiktoken>=0.7.0
orjson>=3.10.7

